+++ braintree/merchant_account_gateway.py
@@ -9,8 +9,12 @@ class MerchantAccountGateway(object):
         self.config = gateway.config
 
     def create(self, params={}):
-        Resource.verify_keys(params, MerchantAccount.create_signature())
+        Resource.verify_keys(params, MerchantAccountGateway._detect_signature(params))
         return self._post("/merchant_accounts/create_via_api", {"merchant_account": params})
+
+    def update(self, merchant_account_id, params={}):
+        Resource.verify_keys(params, MerchantAccountGateway._update_signature())
+        return self._put("/merchant_accounts/%s/update_via_api" % merchant_account_id, {"merchant_account": params})
 
     def _post(self, url, params={}):
         response = self.config.http().post(url, params)
@@ -18,3 +22,125 @@ class MerchantAccountGateway(object):
             return SuccessfulResult({"merchant_account": MerchantAccount(self.gateway, response["merchant_account"])})
         elif "api_error_response" in response:
             return ErrorResult(self.gateway, response["api_error_response"])
+
+    def _put(self, url, params={}):
+        response = self.config.http().put(url, params)
+        if "merchant_account" in response:
+            return SuccessfulResult({"merchant_account": MerchantAccount(self.gateway, response["merchant_account"])})
+        elif "api_error_response" in response:
+            return ErrorResult(self.gateway, response["api_error_response"])
+
+    @staticmethod
+    def _detect_signature(attributes):
+        if attributes.has_key('applicant_details'):
+            # Warn deprecated
+            return MerchantAccountGateway._create_deprecated_signature()
+        else:
+            return MerchantAccountGateway._create_signature()
+
+    @staticmethod
+    def _create_deprecated_signature():
+        return [
+            {'applicant_details': [
+                'company_name',
+                'first_name',
+                'last_name',
+                'email',
+                'phone',
+                'date_of_birth',
+                'ssn',
+                'tax_id',
+                'routing_number',
+                'account_number',
+                {'address': [
+                    'street_address',
+                    'postal_code',
+                    'locality',
+                    'region']}
+                ]
+            },
+            'tos_accepted',
+            'master_merchant_account_id',
+            'id'
+        ]
+
+    @staticmethod
+    def _create_signature():
+        return [
+            {'individual': [
+                'first_name',
+                'last_name',
+                'email',
+                'phone',
+                'date_of_birth',
+                'ssn',
+                {'address': [
+                    'street_address',
+                    'postal_code',
+                    'locality',
+                    'region']}
+                ]
+            },
+            {'business': [
+                'dba_name',
+                'legal_name',
+                'tax_id',
+                {'address': [
+                    'street_address',
+                    'postal_code',
+                    'locality',
+                    'region']}
+                ]
+            },
+            {'funding': [
+                'routing_number',
+                'account_number',
+                'destination',
+                'email',
+                'mobile_phone',
+                ]
+            },
+            'tos_accepted',
+            'master_merchant_account_id',
+            'id'
+        ]
+
+    @staticmethod
+    def _update_signature():
+        return [
+            {'individual': [
+                'first_name',
+                'last_name',
+                'email',
+                'phone',
+                'date_of_birth',
+                'ssn',
+                {'address': [
+                    'street_address',
+                    'postal_code',
+                    'locality',
+                    'region']}
+                ]
+            },
+            {'business': [
+                'dba_name',
+                'legal_name',
+                'tax_id',
+                {'address': [
+                    'street_address',
+                    'postal_code',
+                    'locality',
+                    'region']}
+                ]
+            },
+            {'funding': [
+                'routing_number',
+                'account_number',
+                'destination',
+                'email',
+                'mobile_phone',
+                ]
+            },
+            'master_merchant_account_id',
+            'id'
+        ]